insert logic of Core/IdeaCollection/Process into Core/Process
