+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Core/Process/Module.ts
@@ -2,7 +2,7 @@ import * as AdhHttpModule from "../Http/Module";
 import * as AdhNamesModule from "../Names/Module";
 import * as AdhTopLevelStateModule from "../TopLevelState/Module";
 
-import * as AdhProcess from "./Process";
+import * as Process from "./Process";
 
 
 export var moduleName = "adhProcess";
@@ -14,11 +14,12 @@ export var register = (angular) => {
             AdhNamesModule.moduleName,
             AdhTopLevelStateModule.moduleName
         ])
-        .provider("adhProcess", AdhProcess.Provider)
+        .provider("adhProcess", Process.Provider)
         .directive("adhWorkflowSwitch", [
-            "adhConfig", "adhHttp", "adhPermissions", "$q", "$translate", "$window", AdhProcess.workflowSwitchDirective])
-        .directive("adhProcessView", ["adhTopLevelState", "adhProcess", "$compile", AdhProcess.processViewDirective])
-        .directive("adhProcessListItem", ["adhConfig", "adhHttp", "adhNames", AdhProcess.listItemDirective])
-        .directive("adhProcessListing", ["adhConfig", AdhProcess.listingDirective])
-        .directive("adhCurrentProcessTitle", ["adhTopLevelState", "adhHttp", AdhProcess.currentProcessTitleDirective]);
+            "adhConfig", "adhHttp", "adhPermissions", "$q", "$translate", "$window", Process.workflowSwitchDirective])
+        .directive("adhProcessView", ["adhTopLevelState", "adhProcess", "$compile", Process.processViewDirective])
+        .directive("adhProcessListItem", ["adhConfig", "adhHttp", "adhNames", Process.listItemDirective])
+        .directive("adhProcessListing", ["adhConfig", Process.listingDirective])
+        .directive("adhCurrentProcessTitle", ["adhTopLevelState", "adhHttp", Process.currentProcessTitleDirective])
+        .directive("adhIdeaCollectionDetail", ["adhConfig", "adhEmbed", "adhHttp", "adhPermissions", "$q", Process.detailDirective]);
 };
